feat: enhance accessibility and image handling
Add viewport and default description meta tags, accessible skip link, and improved navigation semantics. Introduce responsive image shortcode and layout container with basic typographic rhythm.
