+++ src/assets/css/main.css
@@ -7,4 +7,19 @@
   button:focus-visible {
     @apply outline-none ring-2 ring-offset-2 ring-blue-600;
   }
+
+  .skip-link {
+    @apply sr-only;
+  }
+  .skip-link:focus {
+    @apply not-sr-only absolute top-0 left-0 p-2 bg-white text-blue-600;
+  }
+
+  .container {
+    @apply max-w-3xl mx-auto px-4;
+  }
+
+  main > * + * {
+    @apply mt-4;
+  }
 }
